DONE--Points and Lines
Draw points and then use lines to make them connecting.
You can use space to see all steps and the different function of drawing
lines
There are three function to draw different lines and one function to
draw points.
The points here is crrated randomly and the seed of random is current
time.